Deflake WPT test registration-schedule-job.https.html
Make it so register jobs are not coalesced if they have
different updateViaCache properties, as per spec:
https://w3c.github.io/ServiceWorker/#dfn-job-equivalent

This makes the test about updateViaCache pass. The test was also
previously flaky in Chromium because it failed in two different ways
after coalescing the jobs: either registration.installing is null or
it's non-null and has the wrong updateViaCache property.

Other browsers also seemed to fail the test because they do not
make a new job and were possibly flaky. This CL also rearranges the
test to help browsers fail in a more stable way by accounting for
registration.installing being null. However it can still be flaky if
one register job is able to finish before the other one is queued up,
so it doesn't have a chance to coalesced.

Bug: 979593
Change-Id: I5395019d5733a6365b0b0ae2ae6051a077fb8176