[Reland]: Repaint when bgcolor animation is triggered
This CL is the reland of:
https://chromium-review.googlesource.com/c/chromium/src/+/2698847

The above CL was reverted because the layout test added was flaky.
In this CL, PS#1 is the reverted CL, and we need to make sure
that the test is no longer flaky.

It turns out that it was the non-virtual version that was flaky,
and the reason is that double raf doesn't seem to be enough to
make the animation jump into the mid-point. I have tested
locally with 500 runs, and the failure rate is ~99%, Oh, that's
bad...

The reason is that run_web_tests might run with a 1ms raf timer
which would be too fast to guarantee the screenshot would be
after the step change.

The solution is very simple, which is to use triple raf instead
of double raf, I have ran both the virtual and non-virtual version
of the test, each one for 500 times locally, and there is no
failure at all.

Bug: 1177320
Change-Id: I95f576994a5d67bf8319b5b1f33e8e6fca6d4bb8